Codec.Compression.Snappy

Description

This module provides fast, pure Haskell bindings to Google's Snappy compression and decompression library: http://code.google.com/p/snappy/

These functions operate on strict bytestrings, and thus use as much memory as both the entire compressed and uncompressed data.

Documentation

compress :: ByteString -> ByteString

Compress data into the Snappy format.

decompress :: ByteString -> ByteString

Decompress data in the Snappy format.

If the input is not compressed or is corrupt, an exception will be thrown.